+++ CHANGELOG.rst
@@ -135,11 +135,20 @@ Please update your code accordingly:
 Old changes (version < 1.x)
 ===========================
 
+2025-01-04
+----------
+**version 0.5.4**
+
+* (python) removed Python 3.8 from the supported versions,
+* (debug) fixed `DivisionByZeroWarning` when semivariogram range is equal to 0
+
 2024-10-26
 ----------
 **version 0.5.3**
 
 * (logic) debugged variance error calculations for Area-to-Point Poisson Kriging
+* (bug) https://github.com/DataverseLabs/pyinterpolate/issues/428
+* (enhancement) added universal kriging functionality along with multivariate regression
 
 2024-06-26
 ----------